% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_changed_formats.R
\name{add_changed_formats}
\alias{add_changed_formats}
\title{Format \code{openxlsx} worksheet based on changes}
\usage{
add_changed_formats(
  wb,
  cur.sheet,
  sheet.comp,
  rows.invert = NULL,
  cols.invert = NULL,
  df.invert = NULL,
  nofillStyle = NULL,
  changeStyle = NULL,
  posStyle = NULL,
  negStyle = NULL
)
}
\arguments{
\item{wb}{openxlsx workbook to make on which to make changes.}

\item{cur.sheet}{sheet name in openxlsx workbook on which to make changes.}

\item{sheet.comp}{list of comparison dataframes generated by \code{sheet_comp()}}

\item{rows.invert}{Optional vector of row numbers to invert color scheme for increase vs decrease.}

\item{cols.invert}{Optional vector of column numbers to invert color scheme for increase vs decrease.}

\item{df.invert}{Optional data frame with \verb{$row} and \verb{$col} entries to indentify individual cells to invert color schemes for increase vs decrease. Defaults to NULL.}

\item{nofillStyle}{Optional openxlsx style object for cells with no changes flagged. Default has black text, white foreground. Create custom style with \code{openxlsx::createStyle()}.}

\item{changeStyle}{As \code{nofillStyle}, but for non-numeric cells with changed values. Default has black text, light purple forecround.}

\item{posStyle}{As \code{nofillStyle}, but for numeric cells that increase in value. Default has black text, light coral foreground.}

\item{negStyle}{As \code{nofillStyle}, but for numeric cells that decrease in value. Default has black text, light green foreground.}
}
\description{
Highlights cells that changed, coloring differently for increasing, decreasing, and non-numeric cells.
Typically used on a worksheet that contains the \verb{$sheet.diff} dataframe from the same sheet comparison as provided
in the \code{cur.sheet} argument. In some cases it may be useful to define custom color schemes (e.g., if increasing numbers are good
and decreasing numbers are bad, you may want green and red foregrounds for those types of changes). Individual styles can be provided
with optional arguments see \code{?openxlsx::createStyle} for options in defining styles.
}
\details{
In some cases it may make sense to reverse the color scheme of numeric changes for individual rows, columns,
or cells (e.g., when scanning fishery model outputs, increasing fish escapement and decreasing fish exploitation rates logically should both show the same color.
Similarly, increasing profits and decreasing costs logically should both show the same color.). Optional arguments \code{rows.invert}, \code{cols.invert}, and
\code{df.invert} allow you to specify individual regions of the dataframe to reverse the color scheme.
}
