% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_sheets.R
\name{align_rows}
\alias{align_rows}
\title{Re-arrange rows of one matrix or dataframe to maximize similarity to a second.}
\usage{
align_rows(sheet_to_change, sheet_to_match, verbose = FALSE)
}
\arguments{
\item{sheet_to_change}{Dataframe or matrix that should be arranged.}

\item{sheet_to_match}{Dataframe or matrix to compare against. Must have same dimensions as \code{sheet_to_change}.}

\item{verbose}{Warn if re-arranging? Logical, defaults to FALse}
}
\value{
re-arrange version of \code{sheet_to_change}
}
\description{
Re-arranges rows of first input to maximize similarity with second matrix.
\code{sheet_to_change} and \code{sheet_to_match} must have the same dimensions.
}
\details{
Re-arrangement is ordered by calculating similarity using \code{make_alignment_matrix} and then solving this as a linear sum assignment problem using the Hungarian method.
}
\examples{
  ## example problem:
  ## two versions of a datafame that almost
  ##    match aside from the first dataframe containing an extra row of NAs
  mat1 <- mtcars[c(1,1,5),]
  mat1[2, ] = NA
  mat2 <- mtcars[c(1,5), ]
  mat2[1,1] = 10

  res <- pad_sheets(mat1, mat2)
  temp <- align_rows(res[[1]], res[[2]])
  temp <- align_rows(res[[1]], res[[2]], verbose = TRUE)
}
