% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_range_translate.R
\name{cell_range_translate}
\alias{cell_range_translate}
\title{Translates from excel cell address to rows and columns}
\usage{
cell_range_translate(x, expand = TRUE, start = "A1")
}
\arguments{
\item{x}{Single string of individual cell or cell range (e.g. "D6" or "D6:AC8")}

\item{expand}{If \code{TRUE} (default), provides the row and column for all cells in the range. If \code{False}, provides just the
row and column of the start and end cells.}

\item{start}{Optional argument to account for offset when matching cells in an excel file to a dataframe
when the dataframe was generated by reading the excel file starting at a location other than "A1".}
}
\value{
dataframe of addresses for each cell in the range, where \verb{$row} gives the row number and \verb{$col} gives the column number.
}
\description{
Translates from excel cell address to rows and columns
}
\examples{
cell_range_translate("D6")
cell_range_translate("A2:H3")
}
