% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_stylize.R
\name{cells_stylize}
\alias{cells_stylize}
\title{Apply style to worksheet based on one or more excel-style cell ranges}
\usage{
cells_stylize(wb, sheet, style, block.ranges, stack = TRUE)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{sheet}{character corresponding to sheet name of openxlsx workbook object \code{wb}.}

\item{style}{\code{openxlsx} cell style, created with \code{openxlsx::createStyle()}. This can include text size, bolding or italics, text wrapping, foreground color, text color, etc. See \code{?openxlsx::createStyle} for details.}

\item{block.ranges}{One or more cell ranges specified in excel format (e.g. \verb{c("A1:D5, "B6", "A8:D8")})}

\item{stack}{Should style be appended to existing styles (\code{TRUE}) or replace existing styles (\code{FALSE}). Defaults to \code{TRUE}.}
}
\description{
Apply style to worksheet based on one or more excel-style cell ranges
}
