% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_handling.R
\name{compare_lists}
\alias{compare_lists}
\title{Compare all pairwise combinations of list items for two lists of matching dimensions}
\usage{
compare_lists(list_1, list_2)
}
\arguments{
\item{list_1}{list of items}

\item{list_2}{list of items, same dimensions as \code{list_1}}
}
\description{
Intended to handle lists of vectors with matching dimensions; typically these will be generated from
\code{\link[=listify_sheet]{listify_sheet()}}. Returns a matrix where each value is the proportion of items that match.
}
