% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_handling.R
\name{listify_sheet}
\alias{listify_sheet}
\title{Converts dataframe or matrix into list}
\usage{
listify_sheet(df, direction, digits = 10)
}
\arguments{
\item{df}{dataframe or matrix}

\item{direction}{Should the function compare row-wise (1) or column-wise (2)}

\item{digits}{How many digits should numeric values be rounded to before comparing? Numeric, defaults to 10.}
}
\value{
list, where each item is a row (\code{direction = 1}) or a column (\code{direction = 2}) of argument \code{df}, with the values transformed using \code{\link[=make_cell_comparable]{make_cell_comparable()}}.
}
\description{
Converts dataframe or matrix into list
}
\examples{
listify_sheet(mtcars, direction = 1)
listify_sheet(mtcars, direction = 2)
}
