% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_handling.R
\name{make_alignment_matrix}
\alias{make_alignment_matrix}
\title{All row-wise or column-wise comparisons of two dataframes}
\usage{
make_alignment_matrix(df_1, df_2, direction, digits = 10)
}
\arguments{
\item{df_1}{First dataframe or matrix to compare}

\item{df_2}{Second dataframe or matrix to compare}

\item{direction}{Should the function compare row-wise (1) or column-wise (2)}

\item{digits}{How many digits should numeric values be rounded to before comparing? Numeric, defaults to 10.}
}
\value{
Matrix of all pairwise comparisons of columns or rows; values are the proportion of cells that match.
}
\description{
Intended for use on objects read in from excel, which may contain formatting complications.
}
\examples{
mat1 = matrix(rnorm(12), 3, 4)

mat2 = mat1

mat2[3,] = mat1[1,]
mat2[1,1] = 10

make_alignment_matrix(mat1, mat2, 2)
}
