% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_handling.R
\name{make_cell_comparable}
\alias{make_cell_comparable}
\title{Processes a character atomic for easier comparisons}
\usage{
make_cell_comparable(x, digits = 10)
}
\arguments{
\item{x}{Atomic to clean. Recommend providing only characters.}

\item{digits}{Number of digits to round to. Defaults to 10.}
}
\value{
Character atomic
}
\description{
Cleans up atomics read in from excel to make for more consistent comparisons.
}
\details{
Excel cells read into R are often read in as characters, and this can lead to funkiness
that might make it hard to compare to another version of the cell read in under funky conditions.
For example, miniscule numeric differences lead to different non-comparable character vectors. Further,
if a cell is formatted as \% in one file and not in another, the raw versions read in from excel will
appear distinct.

If the input is meant to be a character, the function will return a character. If the input is a number
that was read as a character, this function will, as appropriate (a) remove commas, (b) convert from
percent to numeric, and then round to the nearest \code{digits} digits (default: 10). The function will
then return this as a character (for easier application to vectors of mixed contents). If the input is
\code{NA}, this function returns "".
}
\examples{
make_cell_comparable("1,503.999999999999995")
make_cell_comparable(NA)
make_cell_comparable("escapement")
make_cell_comparable("13.00000000000005\%")
}
