% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_to_a1.R
\name{mat_to_a1}
\alias{mat_to_a1}
\title{Provide Excel addresses from matrix of logicals}
\usage{
mat_to_a1(mat)
}
\arguments{
\item{mat}{matrix of logical values}
}
\value{
vector of characters containing addresses in excel "A1" format.
}
\description{
Useful when working with functions that need excel formats (e.g. \code{openxlsx2} functions).
Does not try to simplify blocks of TRUEs into block addresses.
}
\examples{
mat <- matrix(FALSE, ncol = 4, nrow = 3)
mat[1, 1] <- TRUE
mat[2, 3] <- TRUE
mat
mat_to_a1(mat)
}
