% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_sheets.R
\name{pad_sheets}
\alias{pad_sheets}
\title{Add NA rows or columns to either of two dataframes to give same dimensions}
\usage{
pad_sheets(df1, df2)
}
\arguments{
\item{df1}{dataframe or matrix}

\item{df2}{dataframe or matrix}
}
\value{
list, with padded version of \code{df1} as first item and padded version of \code{df2} as second item
}
\description{
Add NA rows or columns to either of two dataframes to give same dimensions
}
\examples{
  mat1 <- mtcars[c(1,1,5),]
  mat2 <- mtcars[c(1,5), ]
  res <- pad_sheets(mat1, mat2)
}
