% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_line_handling.R
\name{score_df_match}
\alias{score_df_match}
\title{Score matching-ness of two dataframes}
\usage{
score_df_match(df_1, df_2, digits = 10)
}
\arguments{
\item{df_1}{First dataframe or matrix to compare}

\item{df_2}{Second dataframe or matrix to compare}

\item{digits}{How many digits should numeric values be rounded to before comparing? Numeric, defaults to 10.}
}
\value{
Single numeric, with average matchedness of the two dataframes.
}
\description{
For use in handling addition/removal of blank rows.
}
\examples{
df2 = df1 = as.data.frame(matrix(1:10, 2, 5))
## currently: perfect match
score_df_match(df1, df2)
## change one value
df2[1,1] = df2[1,1] + .1
score_df_match(df1, df2)
## Does the rounding work right? set digits = 0 to ignore the addition of 0.1 above
score_df_match(df1, df2, digits = 0)
## If the two matrices don't match, should get value of 0
score_df_match(df1, df2+.1)
}
