% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_comp.R
\name{sheet_comp_basic}
\alias{sheet_comp_basic}
\title{Stripped down version of sheet_comp}
\usage{
sheet_comp_basic(t1, t2)
}
\arguments{
\item{t1}{First dataframe}

\item{t2}{Second dataframe, same dimensions as first.}
}
\value{
logical matrix
}
\description{
Stripped down version of sheet_comp
}
