% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_get_col_widths.R
\name{wb_get_col_widths}
\alias{wb_get_col_widths}
\title{Get column widths of openxlsx2 Workbook sheet}
\usage{
wb_get_col_widths(wb, sheet = openxlsx2::current_sheet())
}
\arguments{
\item{wb}{openxlsx2 workbook object}

\item{sheet}{sheet name, defaults to current sheet}
}
\value{
Dataframe with $col (column names) and $width (column widths)
}
\description{
Code by Jan Marvin. This function will likely be superceded by a function in openxlsx2 in the future.
}
\examples{
library(openxlsx2)
wb <- wb_workbook() |>
  wb_add_worksheet("Sheet1") |>
  wb_add_data(x = mtcars) |>
  wb_set_col_widths(cols = 2:4, widths = 12)

current_widths <- wb_get_col_widths(wb, sheet = "Sheet1")
current_widths
}
