% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xldiff-package.R
\docType{package}
\name{xldiff-package}
\alias{xldiff}
\alias{xldiff-package}
\title{xldiff: Compare excel sheets}
\description{
`xldiff` provides tools to compare excel sheets, broadly inspired by "diff"-type functions. Provided functions can read sheets of two excel files and produce a third file that highlights cells that have changed. In the case of numeric changes, the direction of change is highlighted. These tools do not account for structural changes in the sheets (e.g., the addition of a column), but are useful in tracking changed values in tables or parameter files. Utility functions developed to streamline formatting output files are also more broadly useful in programmatically formatting excel files using openxlsx.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://framverse.r-universe.dev/xldiff}
  \item \url{https://github.com/cbedwards-dfw/xldiff}
  \item \url{https://cbedwards-dfw.github.io/xldiff/}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
