% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_changed_formats.R
\name{add_changed_formats}
\alias{add_changed_formats}
\title{Format \code{openxlsx} worksheet based on changes}
\usage{
add_changed_formats(
  wb,
  cur.sheet,
  sheet.comp,
  rows.invert = NULL,
  cols.invert = NULL,
  df.invert = NULL,
  change.color = "#E1C0FF",
  pos.color = "#FFC7CE",
  neg.color = "#C6EFCE"
)
}
\arguments{
\item{wb}{openxlsx workbook to make on which to make changes.}

\item{cur.sheet}{sheet name in openxlsx workbook on which to make changes.}

\item{sheet.comp}{list of comparison dataframes generated by \code{sheet_comp()}}

\item{rows.invert}{Optional vector of row numbers to invert color scheme for increase vs decrease.}

\item{cols.invert}{Optional vector of column numbers to invert color scheme for increase vs decrease.}

\item{df.invert}{Optional data frame with \verb{$row} and \verb{$col} entries to identify individual cells to invert color schemes for increase vs decrease. Defaults to NULL.}

\item{change.color}{Hex code for color to highlight cells that changed (and were not numerics). Defaults to lavendar ("#E1C0FF").}

\item{pos.color}{As \code{change.color}, but for numeric cells that increase in value. Defaults to coral ("#E1C0FF").}

\item{neg.color}{As \code{change.color}, but for numeric cells that decrease in value. Defaults to light green ("#C6EFCE")}
}
\description{
Highlights cells that changed, coloring differently for increasing, decreasing, and non-numeric cells.
Typically used on a worksheet that contains the \verb{$sheet.diff} dataframe from the same sheet comparison as provided
in the \code{cur.sheet} argument. In some cases it may be useful to define custom color schemes (e.g., if increasing numbers are good
and decreasing numbers are bad, you may want green and red foregrounds for those types of changes). Individual styles can be provided
with optional arguments see \code{?openxlsx::createStyle} for options in defining styles.
}
\details{
In some cases it may make sense to reverse the color scheme of numeric changes for individual rows, columns,
or cells (e.g., when scanning fishery model outputs, increasing fish escapement and decreasing fish exploitation rates logically should both show the same color.
Similarly, increasing profits and decreasing costs logically should both show the same color.). Optional arguments \code{rows.invert}, \code{cols.invert}, and
\code{df.invert} allow you to specify individual regions of the dataframe to reverse the color scheme.
}
