% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_diff_table.R
\name{diff_to_table}
\alias{diff_to_table}
\title{Compares two dataframes and returns diff as flextable}
\usage{
diff_to_table(df1, df2, dfnames = NULL, excelify.col.names = FALSE)
}
\arguments{
\item{df1}{first dataframe or tibble}

\item{df2}{second dataframe or tibble}

\item{dfnames}{Character vector with names of the two dataframes to use for table title. Optional.}

\item{excelify.col.names}{Should columns be identified by excel letter convention? Defaults to FALSE. Set to TRUE when comparing dataframes extracted from excel sheets.}
}
\value{
flextable object
}
\description{
Workhorse function for \code{excel_diff_table()}. Can be used to diff any pair of dataframes or tibbles. Diff is applied cell by cell, so results will only be meaningful if the shape of the two objects is the same.
}
\examples{
df1 = mtcars
df2 = mtcars
df2[2,2] = -8
df2[5, c(3,4)] = c(11, 15)
diff_to_table(df1, df2)
}
