% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_diff.R
\name{excel_diff}
\alias{excel_diff}
\title{Minimal spreadsheet comparison function}
\usage{
excel_diff(
  file.1,
  file.2,
  results.name,
  sheet.name,
  extra_format_fun = NULL,
  ...
)
}
\arguments{
\item{file.1}{Filename (including path) for first file to compare}

\item{file.2}{Filename (including path) for second file to compare}

\item{results.name}{Name (including path) for file to save comparison to. Must end in ".xlsx"}

\item{sheet.name}{character string of sheet to compare. If the sheets have different names (or to compare two sheets in one file), can take a character vector of length two, with the sheet names from the first and second file in order.}

\item{extra_format_fun}{Optional function to apply additional formatting, allowing users to specify additional
calls of \code{addStyle()} (or other openxslx functions, like setting column width). First argument must be the workbook
object this function makes changes to; second argument must be the name of the worksheet this function makes
changes to}

\item{...}{Additional arguments passed to \code{extra_format_fun}}
}
\description{
Compares a single sheet between two files, supports providing additional formatting
in the form of the optional \code{extra_format_fun} argument. For more complex use cases (e.g.,
multiple sheet, pre-comparison formatting to compare only specific regions, etc) \code{excel_diff} can
be used as a simple template for writing your own function.
}
\examples{
\dontrun{
filename.1 = "Documents/WDFW FRAM team work/NOF material/NOF 2024/FRAM/Chin1124.xlsx"
filename.2 = "Documents/WDFW FRAM team work/NOF material/NOF 2024/NOF 2/Chin2524.xlsx"

excel_diff(file.1 = filename.1,
          file.2 = filename.2,
          results.name = "Documents/WDFW FRAM team work/NOF material/NOF 2024/test1.xlsx",
          sheet.name = "ER_ESC_Overview_New"
)

## create function to add in some additional formatting:
extra_form_fun = function(wb, sheet){
 ## add bold and increased size for the first two rows.
 openxlsx::addStyle(wb, sheet,
                    style = openxlsx::createStyle(fontSize = 16, textDecoration = "Bold"),
                    rows = 1:2, cols = 1:8, gridExpand = TRUE,
                    stack = TRUE)
 ## add thin inner cell borders
 add_cell_borders(wb, sheet,
                  block.ranges = c("B3:H34") )
 ## add thick outer borders
 add_cell_borders(wb, sheet,
                  block.ranges = c("A2", "B1:D2", "E1:H2",
                                   "A3:A34", "B3:D34", "E3:H34",
                                   "D36:H37"),
                  border.thickness = "medium")
}

excel_diff(file.1 = filename.1,
          file.2 = filename.2,
          results.name = "Documents/WDFW FRAM team work/NOF material/NOF 2024/test2.xlsx",
          sheet.name = "ER_ESC_Overview_New",
          extra_format_fun = extra_form_fun
)
}
}
