% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_diff.R
\name{excel_diff}
\alias{excel_diff}
\title{Minimal spreadsheet comparison function}
\usage{
excel_diff(
  file_1,
  file_2,
  results_name,
  sheet_name,
  sheet_name_file_2 = NULL,
  cell_range = NULL,
  cell_range_2 = NULL,
  proportional_threshold = 0.001,
  absolute_threshold = NULL,
  digits_show = 6,
  realign_rows = TRUE,
  verbose = FALSE,
  extra_width = NULL
)
}
\arguments{
\item{file_1}{Filename (including path) for first file to compare}

\item{file_2}{Filename (including path) for second file to compare}

\item{results_name}{Name (including path) for file to save comparison to. Must end in ".xlsx"}

\item{sheet_name}{Character string of sheet to compare. Can provide vector of character strings to produce comparisons of multiple sheets.}

\item{sheet_name_file_2}{OPTIONAL. Matching sheet names to \code{sheet_name} but for file 2. Use only if the two files have matching sheets with different names. Defaults to NULL.}

\item{cell_range}{OPTIONAL. Character string of excel formatted range (e.g., "A1:D5"); if provided, excel_diff will only compare that range. By default, assumes cell range is the same in both files; if it differs, provide second file's cell range with argument \code{cell_range_2}}

\item{cell_range_2}{OPTIONAL. Like \code{cell_range}; only provide if (a) \code{cell_range} is provided, and (b) the cell range differs between the two files (e.g., the regions to compare are offset in the sheets)}

\item{proportional_threshold}{Sets a threshold of proportional change below which differences should be ignored. For example, a value of 0.1 means any changes less than 10\% will not be flagged as having changed. \code{proportional_threshold} will override this value and behavior if it is provided. Numeric, defaults to 0.001 (0.1\% change).}

\item{absolute_threshold}{Optional. Sets a threshold of absolute change below which differences should be ignored. For example, a value of 0.1 means any changes less than 0.1 will not be flagged as having changed. If provided, will override \code{proportional_threshold}. Numeric, defaults to NULL.}

\item{digits_show}{When there is a change in number values, how many digits should be shown in \verb{## ---> ##}? Numeric, defaults to 6. Recommend not making this so small that flagged changes don't get printed (e.g., if this is 2 and \code{proportional_threshold} is 0.001, 0.1\% changes will get flagged, but only the first two digits will get shown).}

\item{realign_rows}{Should function check if re-arranging row order of sheet in file_1 improves matchingness? Useful if one or more blank rows have been added/removed. If TRUE, row arrangement will match / align with file_2.}

\item{verbose}{Should sheet names be listed as they are diffed? Logical, defaults to TRUE}

\item{extra_width}{How much extra width should be added to columns that changed? Helpful to improve readability, since changed cells have longer entries. Numeric, defaults to 0.4.}
}
\value{
Invisibly returns the number of cells that differed
}
\description{
Compares a single sheet between two files, creating a new file that uses the formatting of
the existing sheets (except highlighting) and highlights cells that differed.
Small numeric differences between cell values can be the result of happenstance ("decimal dust") - the \code{digits_signif} and \code{proportional.diff} arguments can control behavior to ignore minor numeric differences.
}
\examples{
\dontrun{
filename_1 <- "Chin1124.xlsx"
filename_2 <- "Chin2524.xlsx"

excel_diff(
  file_1 = filename_1,
  file_2 = filename_2,
  results.name = "Chin1124 vs Chin 2524.xlsx",
  sheet_name = "ER_ESC_Overview_New"
)
}
}
\seealso{
\code{\link[=excel_diff_table]{excel_diff_table()}}, \code{\link[=excel_diff_tibble]{excel_diff_tibble()}}
}
