% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_diff_table.R
\name{excel_diff_table}
\alias{excel_diff_table}
\title{Diff excel sheets and show table}
\usage{
excel_diff_table(file.1, file.2, sheet.name)
}
\arguments{
\item{file.1}{Filename (including path) for first file to compare}

\item{file.2}{Filename (including path) for second file to compare}

\item{sheet.name}{character string of sheet to compare. If the sheets have different names (or to compare two sheets in one file), can take a character vector of length two, with the sheet names from the first and second file in order.}
}
\value{
flextable object.
}
\description{
Similar to excel_diff, but returns flextable that can be displayed and navigated in the Rstudio viewer. Only shows rows that have changed. For every row with changes, provides a row of before and after, highlighting changed vlaues (red for the value in file.1, green for the value in file.2)."ROWS" column identifies excel row number and columns identify excel column names.
}
\examples{
\dontrun{
excel_diff_table(file.1 = "C:/Repos/test file 1.xlsx",
file.2 = "C:/Repos/test file 2.xlsx",
sheet = "Sheet1")
}
}
