% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_status_diff.R
\name{formula_status_diff}
\alias{formula_status_diff}
\title{Compare two sheets to see if formula use changed}
\usage{
formula_status_diff(file_1, file_2, sheet_name)
}
\arguments{
\item{file_1}{Filename (including path) for first file to compare}

\item{file_2}{Filename (including path) for second file to compare}

\item{sheet_name}{Character string of a single excel sheet to compare between the files. (Unlike \code{excel_diff}, only one sheet can be compared at a time)}
}
\value{
Nothing
}
\description{
Identifies if one or more cells changed from having a formula to not between two excel sheets.
}
\examples{
\dontrun{
formula_status_diff(
  file_1 = "CohoPugetSoundTAMMInputTemplate2025.xlsx",
  file_2 = "CohoPugetSoundTAMMInputs 2025 v1.5.xlsx",
  sheet_name = "Input_Harvestnew"
)
}
}
